refactor(runtime): share loopback URL helpers; fix stale startup test (BRO-1558)
Addresses Greptile review on the loopback base-URL port fix:

1. Update the stale startup test that asserted the old (buggy) behavior —
   it expected an external `my-host.ts.net:3100` base URL to be rewritten to
   the internal listen port. External hosts now correctly keep their advertised
   port, so the test asserts preservation.

2. Consolidate the duplicated `isLoopbackHost` / `rewriteLocalUrlPort` helpers
   into a single shared module (server/src/url-utils.ts), imported by both
   startup (index.ts) and worktree config (worktree-config.ts). This removes the
   drift risk of two copies and gives one tested source of truth.

Tests moved to url-utils.test.ts (exercises the shared helpers directly).
Verified: url-utils.test.ts + the updated startup assertion pass; tsc --noEmit
clean (exit 0).
