+++ server/src/worktree-config.ts
@@ -3,6 +3,7 @@ import os from "node:os";
 import path from "node:path";
 import type { PaperclipConfig } from "@paperclipai/shared";
 import { resolvePaperclipConfigPath, resolvePaperclipEnvPath } from "./paths.js";
+import { rewriteLocalUrlPort } from "./url-utils.js";
 
 function nonEmpty(value: string | null | undefined): string | null {
   return typeof value === "string" && value.trim().length > 0 ? value.trim() : null;
@@ -25,30 +26,6 @@ function sanitizeWorktreeInstanceId(rawValue: string): string {
     .replace(/-+/g, "-")
     .replace(/^[-_]+|[-_]+$/g, "");
   return normalized || "worktree";
-}
-
-export function isLoopbackHost(host: string): boolean {
-  // Strip surrounding brackets so a URL hostname form ("[::1]") matches too.
-  const normalized = host.trim().toLowerCase().replace(/^\[|\]$/g, "");
-  return normalized === "127.0.0.1" || normalized === "localhost" || normalized === "::1";
-}
-
-export function rewriteLocalUrlPort(rawUrl: string | undefined, port: number): string | undefined {
-  if (!rawUrl) return undefined;
-  try {
-    const parsed = new URL(rawUrl);
-    // The URL API normalizes default ports like :80/:443 to "", so treat them as stable URLs.
-    if (!parsed.port) return rawUrl;
-    // Only rewrite the port for loopback hosts. An explicit external base URL (e.g. a
-    // Tailscale Serve listener on a non-default port like :8443) must survive untouched:
-    // rewriting its port to the internal listen port yields an unreachable URL (scheme/port
-    // mismatch) that then propagates to spawned agents as a dead PAPERCLIP_API_URL. (BRO-1558)
-    if (!isLoopbackHost(parsed.hostname)) return rawUrl;
-    parsed.port = String(port);
-    return parsed.toString();
-  } catch {
-    return rawUrl;
-  }
 }
 
 function parseEnvFile(contents: string): Record<string, string> {
